# backend/admin_openapi.yaml
openapi: 3.0.3
info:
  title: 智糖小助手 - 后台管理API
  description: '智糖小助手后台管理接口文档


    包含用户管理、FAQ管理、标签管理、Prompt管理、消息记录管理等模块。'
  version: 2.0.0
  contact:
    name: 智糖团队
servers:
- url: http://127.0.0.1:8900
  description: 本地开发环境
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Token认证，格式: Bearer <token>'
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: 错误信息
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 操作成功
    User:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        username:
          type: string
          example: user123
        nickname:
          type: string
          example: 用户昵称
        phone_number:
          type: string
          example: '13800138000'
        email:
          type: string
          example: user@example.com
        date_of_birth:
          type: string
          format: date
          example: '1990-01-01'
        is_active:
          type: boolean
          example: true
        is_admin:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            users:
              type: array
              items:
                $ref: '#/components/schemas/User'
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
            total_pages:
              type: integer
...
paths:
...
  /api/users:
    get:
      tags:
      - 用户管理
      summary: 获取用户列表（分页）
      description: 获取用户列表，支持分页、搜索和筛选
      operationId: getUsersList
      parameters:
      - name: page
        in: query
        description: 页码
        required: false
        schema:
          type: integer
          default: 1
      - name: page_size
        in: query
        description: 每页数量
        required: false
        schema:
          type: integer
          default: 20
      - name: keyword
        in: query
        description: 搜索关键词（用户名/昵称/手机号）
        required: false
        schema:
          type: string
      - name: is_active
        in: query
        description: 是否激活
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - 用户管理
      summary: 创建用户
      description: 创建新用户
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - username
              - password
              properties:
                username:
                  type: string
                  example: newuser
                password:
                  type: string
                  example: password123
                nickname:
                  type: string
                  example: 新用户
                phone_number:
                  type: string
                  example: '13800138000'
                email:
                  type: string
                  example: user@example.com
                is_admin:
                  type: boolean
                  example: false
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      user_id:
                        type: integer
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{user_id}:
    get:
      tags:
      - 用户管理
      summary: 获取用户详情
      description: 获取指定用户的详细信息
      operationId: getUserDetail
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
        description: 用户ID
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
      - 用户管理
      summary: 更新用户信息
      description: 更新指定用户的信息
      operationId: updateUser
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
        description: 用户ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                phone_number:
                  type: string
                email:
                  type: string
                date_of_birth:
                  type: string
                  format: date
                is_admin:
                  type: boolean
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - 用户管理
      summary: 删除用户
      description: 删除指定用户
      operationId: deleteUser
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
        description: 用户ID
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{user_id}/status:
    put:
      tags:
      - 用户管理
      summary: 切换用户状态
      description: 切换用户的激活状态
      operationId: toggleUserStatus
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
        description: 用户ID
      responses:
        '200':
          description: 切换成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      is_active:
                        type: boolean
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未授权，需要登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
security:
- BearerAuth: []